Algorithms Deciphering Code
Write a program to find the distribution of all k-tuples in a corpus of English text. The choice of k should be left to the user, but should be less than, say, 10. For simplicity, remove all punctuation and spaces, and convert all letters to lower case. As a test case, apply this to Captain Kidd’s text to see the best way to decipher it. Keep in mind that not all k-tuples will necessarily appear in the corpus.